Модуляция файлов проекта
